requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
email_validator
retry

python-dotenv
redis

bcrypt
cryptography
sqlalchemy

psycopg2-binary

PyMySQL

requests
ratelimit

pillow
